Add Delete option for users (#71)

+++ client/src/components/complain/Complain.jsx
@@ -28,6 +28,11 @@ export default function Complain({ complain, currUser }) {
     }
 
     async function handleDelete() {
+        if (!currUser) {
+            navigate('/login', { state: { redirect: '/complaints' } });
+            return;
+        }
+
         try {
             await api.delete(`/complains/${complain?.id}`);
             window.location.reload();
